+++ src/capi/azext_capi/helpers/binary.py
@@ -8,6 +8,7 @@
 import os
 import platform
 import stat
+import tarfile
 
 from azure.cli.core.azclierror import FileOperationError
 from azure.cli.core.azclierror import InvalidArgumentValueError
@@ -42,6 +43,10 @@ def check_clusterctl(cmd, install=False, install_path=None):
     check_binary(cmd, "clusterctl", install_clusterctl, install=install, install_path=install_path)
 
 
+def check_helm(cmd, install=False, install_path=None):
+    check_binary(cmd, "helm", install_helm, install=install, install_path=install_path)
+
+
 def check_kind(cmd, install=False, install_path=None):
     check_prereq_docker()
     check_binary(cmd, "kind", install_kind, install=install, install_path=install_path)
@@ -66,6 +71,20 @@ def check_binary(cmd, binary_name, install_binary_method, install=False, install
                 install_binary_method(cmd, install_location=install_path)
 
 
+def get_arch(arch=None):
+    """Normalize python's platform.machine() output to match build architectures."""
+    if arch is None:
+        arch = platform.machine()
+    arch = arch.lower()
+    if arch == "x86_64":
+        return "amd64"
+    if arch == "aarch64":
+        return "arm64"
+    if arch == "armv7l":
+        return "arm"
+    return arch
+
+
 def install_clusterctl(_cmd, client_version="latest", install_location=None, source_url=None):
     """
     Install clusterctl, a command-line interface for Cluster API Kubernetes clusters.
@@ -77,12 +96,12 @@ def install_clusterctl(_cmd, client_version="latest", install_location=None, sou
 
     if client_version != "latest":
         source_url += "tags/"
-    source_url += "{}/download/clusterctl-{}-amd64"
+    source_url += "{}/download/clusterctl-{}-{}"
 
     file_url = ""
     system = platform.system()
     if system in ("Darwin", "Linux"):
-        file_url = source_url.format(client_version, system.lower())
+        file_url = source_url.format(client_version, system.lower(), get_arch())
     else:  # TODO: support Windows someday?
         raise ValidationError(f'The clusterctl binary is not available for "{system}"')
 
@@ -100,13 +119,49 @@ def install_clusterctl(_cmd, client_version="latest", install_location=None, sou
     return download_binary(install_location, install_dir, file_url, system, cli)
 
 
-def install_kind(_cmd, client_version="v0.10.0", install_location=None, source_url=None):
+def install_helm(_cmd, client_version="v3.10.2", install_location=None, source_url=None):
+    """
+    Install Helm, an installer and manager for Kubernetes resources.
+    """
+
+    tag = client_version
+    system = platform.system()
+    platform_os = system.lower()
+    arch = get_arch()
+
+    if not source_url:
+        source_url = f"https://get.helm.sh/helm-{tag}-{platform_os}-{arch}.tar.gz"
+
+    # ensure installation directory exists
+    if install_location is not None:
+        install_location = f'{install_location}/helm'
+    else:
+        install_location = _get_default_install_location("helm")
+    install_dir, cli = os.path.dirname(install_location), os.path.basename(install_location)
+    if not os.path.exists(install_dir):
+        os.makedirs(install_dir)
+
+    tarball = f"{install_location}.tar.gz"
+    if download_binary(tarball, install_dir, source_url, system, cli):
+        with tarfile.open(tarball, "r:gz") as tar:
+            for m in tar.getmembers():
+                if m.isreg() and m.name.endswith("helm") or m.name.endswith("helm.exe"):
+                    m.name = os.path.basename(m.name)
+                    tar.extract(m, install_dir)
+                    break
+        os.remove(tarball)
+
+
+def install_kind(_cmd, client_version="v0.17.0", install_location=None, source_url=None):
     """
     Install kind, a container-based Kubernetes environment for development and testing.
     """
 
+    system = platform.system()
+    platform_os = system.lower()
+    arch = get_arch()
     if not source_url:
-        source_url = "https://kind.sigs.k8s.io/dl/{}/kind-{}-amd64"
+        source_url = f"https://kind.sigs.k8s.io/dl/{client_version}/kind-{platform_os}-{arch}"
 
     # ensure installation directory exists
     if install_location is not None:
@@ -119,18 +174,7 @@ def install_kind(_cmd, client_version="v0.10.0", install_location=None, source_u
     if not os.path.exists(install_dir):
         os.makedirs(install_dir)
 
-    file_url = ""
-    system = platform.system()
-    if system == "Windows":
-        file_url = source_url.format(client_version, "windows")
-    elif system == "Linux":
-        file_url = source_url.format(client_version, "linux")
-    elif system == "Darwin":
-        file_url = source_url.format(client_version, "darwin")
-    else:
-        raise InvalidArgumentValueError(f'System "{system}" is not supported by kind.')
-
-    return download_binary(install_location, install_dir, file_url, system, cli)
+    return download_binary(install_location, install_dir, source_url, system, cli)
 
 
 def install_kubectl(cmd, client_version="latest", install_location=None, source_url=None):
@@ -152,7 +196,7 @@ def install_kubectl(cmd, client_version="latest", install_location=None, source_
 
     file_url = ""
     system = platform.system()
-    base_url = source_url + "/{}/bin/{}/amd64/{}"
+    base_url = source_url + "/{}/bin/{}/{}/{}"
 
     # ensure installation directory exists
     if install_location is not None:
@@ -165,13 +209,13 @@ def install_kubectl(cmd, client_version="latest", install_location=None, source_
     if not os.path.exists(install_dir):
         os.makedirs(install_dir)
 
+    arch = get_arch()
     if system == "Windows":
-        file_url = base_url.format(client_version, "windows", "kubectl.exe")
+        file_url = base_url.format(client_version, "windows", arch, "kubectl.exe")
     elif system == "Linux":
-        # TODO: Support ARM CPU here
-        file_url = base_url.format(client_version, "linux", "kubectl")
+        file_url = base_url.format(client_version, "linux", arch, "kubectl")
     elif system == "Darwin":
-        file_url = base_url.format(client_version, "darwin", "kubectl")
+        file_url = base_url.format(client_version, "darwin", arch, "kubectl")
     else:
         raise InvalidArgumentValueError(
             f"Proxy server ({system}) does not exist on the cluster."