+++ src/capi/HISTORY.rst
@@ -3,6 +3,18 @@
 Release History
 ===============
 
+0.1.1
+++++++
+
+* Apply Windows kube-proxy manifest
+
+0.1.0
+++++++
+
+* Default `az capi create --kubernetes-version` to v1.25.3
+* Install Calico CNI with official Helm chart
+* Add "--wait-for-nodes" argument to "az capi create"
+
 0.0.6
 ++++++